refactor(ClusterInfo): move to containers
