+++ src/containers/ClusterInfo/ClusterInfo.scss
